src/requirements.txt
alembic==1.17.1
azure-core==1.36.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
azure-storage-blob==12.27.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
Flask==2.3.3
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
opencensus-ext-flask==0.8.2
opencensus-ext-logging==0.1.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyodbc==5.3.0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
